Updating the forked repository

+++ tests/scalapack/scalapack_06b.cc
@@ -1,6 +1,6 @@
 // ---------------------------------------------------------------------
 //
-// Copyright (C) 2017 - 2018 by the deal.II authors
+// Copyright (C) 2017 - 2019 by the deal.II authors
 //
 // This file is part of the deal.II library.
 //
@@ -105,7 +105,7 @@ test(const unsigned int size,
          &*work.begin(),
          &lwork,
          &info);
-    lwork = work[0];
+    lwork = static_cast<int>(work[0]);
     work.resize(lwork);
     syev(&jobz,
          &uplo,
